+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # create-trigger
 
+## 3.0.0-beta.12
+
+### Patch Changes
+
+- @trigger.dev/core@3.0.0-beta.12
+- @trigger.dev/yalt@3.0.0-beta.12
+
 ## 3.0.0-beta.11
 
 ### Patch Changes